6-size.c

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(char));
-printf("size of an int: %zu bytes(s)\n", (unsigned long)sizeof(int));
-printf("size of a long int: %zu bytes(s)\n", (unsigned long)sizeof(long int));
-printf("size of a long long int: %zu bytes(s)\n", (unsigned long)sizeof(long long int));
-printf("size of a float: %zu bytes(s)\n", (unsigned long)sizeof(float));
+printf("Size of a char: %zu byte(s)\n", sizeof(char));
+printf("size of an int: %zu bytes(s)\n", sizeof(int));
+printf("size of a long int: %zu bytes(s)\n", sizeof(long int));
+printf("size of a long long int: %zu bytes(s)\n", sizeof(long long int));
+printf("size of a float: %zu bytes(s)\n", sizeof(float));
 return (0);
 }
